% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpredgi.R
\name{mlpredgi}
\alias{mlpredgi}
\title{Building a Function with different Machine Learning algorithms}
\usage{
mlpredgi(pred, data, exclude)
}
\arguments{
\item{pred}{as the variable to predict}

\item{data}{as the dataset which contains the variables}

\item{exclude}{as optional variables to exclude in the model}
}
\value{
A table with the summary of different models results
}
\description{
This function runs different machine learning algorithms to predict
gene intereaction based on the other variables of the dataset.
}
\details{
This function compares different regression models like Linear
Regression and some machine learning algorithms: K-NN,
Artificial Neural Networks, Support Vector Machine, Decision
Tree and Random Forest, and shows the RMSE (Root Mean Square Error)
of them to compare the performance with different parameters.

The function splits also the dataset in training and test with a
proportion of 67-33/100 sample rows, to make the training and
validation of each machine learning algorithm.
The function allows also the possibility to exclude one or some
variables in the model.
}
\examples{
mlpredgi("SGAsco", SGA10_full, exclude="NPC")
mlpredgi("SGAsco", SGA10_full, exclude=c("PCC", "fcMF"))

}
\seealso{
\code{lm}

\code{train}

\code{neuralnet}

\code{ksvm}

\code{rpart}

\code{randomForest}
}
\author{
Jordi Cabral
}
