% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SGA16_full}
\alias{SGA16_full}
\title{Different variables of 635 genes to compute gene interaction}
\format{A data frame with 635 rows and 15 variables:

\describe{
     \item{SGAscore}{Synthetic Genetic Array score, a value between
     -1 and 0}
     \item{PCC}{Pearson Correlation Coefficient}
     \item{NPC}{Networked Partial Correlation}
     \item{fcXX}{Gene Ontology functional information}
     ...
}}
\source{
\url{https://science.sciencemag.org/content/353/6306/aaf1420}

\url{https://repositori.upf.edu/handle/10230/32383}

\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC524818/}
}
\usage{
SGA16_full
}
\description{
An example dataset containing 15 variables of 635 genes
}
\keyword{datasets}
